+++ cg/meta/upload/gisaid/gisaid.py
@@ -2,9 +2,10 @@
 
 import logging
 from pathlib import Path
-from typing import List
+from typing import List, Optional
 
 from cg.apps.housekeeper.hk import HousekeeperAPI
+from cg.apps.lims import LimsAPI
 from cg.models.cg_config import CGConfig
 from cg.store import models, Store
 from cg.utils import Process
@@ -23,49 +24,55 @@ class GisaidAPI:
 
     def __init__(self, config: CGConfig):
         self.housekeeper_api: HousekeeperAPI = config.housekeeper_api
+        self.lims_api: LimsAPI = config.lims_api
         self.status_db: Store = config.status_db
         self.gisaid_submitter: str = config.gisaid.submitter
         self.gisaid_binary: str = config.gisaid.binary_path
         self.process = Process(binary=self.gisaid_binary)
 
-    def get_fasta_sequence(self, fastq_path: str) -> str:
-        """Get header from fasta file. Assuming un zipped file"""
+    def get_new_header(self, samples: List[GisaidSample], old_header: str) -> str:
+        for sample in samples:
+            sample_id = sample.covv_subm_sample_id
+            if old_header.find(sample_id) != -1:
+                return f">{sample.covv_virus_name}\n"
+        raise FastaSequenceMissingError
 
-        with open(fastq_path) as handle:
-            fasta_lines = handle.readlines()
-            try:
-                return fasta_lines[1].rstrip("\n")
-            except:
-                raise FastaSequenceMissingError
-
-    def get_gisaid_fasta_objects(self, gsaid_samples: List[GisaidSample]) -> List[FastaFile]:
+    def get_gisaid_fasta(
+        self, gsaid_samples: List[GisaidSample], family_id: str
+    ) -> List[FastaFile]:
         """Fetch a fasta files form house keeper for batch upload to gisaid"""
 
-        fasta_objects = []
-        for sample in gsaid_samples:
-            hk_version: hk_models.Version = self.housekeeper_api.last_version(
-                bundle=sample.family_id
-            )
-            if not hk_version:
-                LOG.info("Family ID: %s not found in hose keeper", sample.family_id)
-                raise HousekeeperVersionMissingError
-            fasta_file: str = self.housekeeper_api.files(
-                version=hk_version.id, tags=["consensus", sample.cg_lims_id]
-            ).first()
-            fasta_sequence: str = self.get_fasta_sequence(fastq_path=fasta_file.full_path)
-            fasta_obj = FastaFile(header=sample.covv_virus_name, sequence=fasta_sequence)
-            fasta_objects.append(fasta_obj)
-        return fasta_objects
+        hk_version: hk_models.Version = self.housekeeper_api.last_version(bundle=family_id)
+        if not hk_version:
+            LOG.info("Family ID: %s not found in housekeeper", family_id)
+            raise HousekeeperVersionMissingError
+        fasta_file: str = self.housekeeper_api.files(
+            version=hk_version.id, tags=["consensus"]
+        ).first()
+        gisaid_delivery_fasta = []
+        with open("/Users/maya.brandi/opt/395698.consensus.fa") as handle:
+            fasta_lines = handle.readlines()
+            for line in fasta_lines:
+                if line[0] == ">":
+                    gisaid_delivery_fasta.append(
+                        self.get_new_header(old_header=line, samples=gsaid_samples)
+                    )
+                else:
+                    gisaid_delivery_fasta.append(line)
 
-    def build_gisaid_fasta(self, gsaid_samples: List[GisaidSample], file_name: str) -> Path:
+        return gisaid_delivery_fasta
+
+    def build_gisaid_fasta(
+        self, gsaid_samples: List[GisaidSample], file_name: str, family_id: str
+    ) -> Path:
         """Concatenates a list of consensus fastq objects"""
 
         file: Path = Path(file_name)
-        fasta_objects: List[FastaFile] = self.get_gisaid_fasta_objects(gsaid_samples=gsaid_samples)
+        fasta_lines: List[str] = self.get_gisaid_fasta(
+            gsaid_samples=gsaid_samples, family_id=family_id
+        )
         with open(file, "w") as write_file_obj:
-            for fasta_object in fasta_objects:
-                write_file_obj.write(f">{fasta_object.header}\n")
-                write_file_obj.write(f"{fasta_object.sequence}\n")
+            write_file_obj.writelines(fasta_lines)
         return file
 
     def get_sample_row(self, gisaid_sample: GisaidSample) -> List[str]:
@@ -101,14 +108,26 @@ class GisaidAPI:
         samples: List[models.Sample] = self.status_db.get_sequenced_samples(family_id=family_id)
         gisaid_samples = []
         for sample in samples:
+            sample_id: str = sample.internal_id
             gisaid_sample = GisaidSample(
                 family_id=family_id,
-                cg_lims_id=sample.internal_id,
+                cg_lims_id=sample_id,
                 covv_subm_sample_id=sample.name,
                 submitter=self.gisaid_submitter,
                 fn=f"{family_id}.fasta",
-                covv_collection_date="2020-11-22",  # sample.collection_date
-                lab="Stockholm",  # sample.originating_lab,
+                covv_collection_date=self.lims_api.get_sample_attribute(
+                    lims_id=sample_id, key="collection_date"
+                ),
+                region=self.lims_api.get_sample_attribute(lims_id=sample_id, key="region"),
+                region_code=self.lims_api.get_sample_attribute(
+                    lims_id=sample_id, key="region_code"
+                ),
+                covv_orig_lab=self.lims_api.get_sample_attribute(
+                    lims_id=sample_id, key="original_lab"
+                ),
+                covv_orig_lab_addr=self.lims_api.get_sample_attribute(
+                    lims_id=sample_id, key="original_lab_address"
+                ),
             )
             gisaid_samples.append(gisaid_sample)
         return gisaid_samples